+++ apps/kyb-app/CHANGELOG.md
@@ -1,5 +1,14 @@
 # kyb-app
 
+## 0.3.161
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.88
+  - @ballerine/workflow-browser-sdk@0.6.110
+
 ## 0.3.160
 
 ### Patch Changes